Fix a few errors

+++ libc/stdio/printf.c
@@ -18,7 +18,7 @@ int printf(const char* restrict format, ...) {
 	va_list parameters;
 	va_start(parameters, format);
 
-	int written = 0;
+	size_t written = 0;
 
 	while (*format != '\0') {
 		size_t maxrem = INT_MAX - written;
